errors fixed! read desc.
added some comments around for reference on what i did to fix it. the code is now reusable: any x-value can be given and any scale factor can be given, and it will generate the two flags. yayyy
p.s. we will need to present it without the comments, bc she will deduct marks if comments are present